requirements.txt
google-adk
fastapi 
langchain 
twilio 
sqlalchemy 
google-auth 
google-api-python-client
decouple
python-dotenv
uvicorn 
langchain_community
psycopg2
openai
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
pandas